requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
distlib==0.3.8
Django==4.2.16
djangorestframework==3.15.2
filelock==3.16.0
mysqlclient==2.2.4
platformdirs==4.3.2
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.4
